Created other Rooms that inhertit from Room, created new trap, SpikeTrap, and new wall type LWall

+++ Final Project/BasicOpenGLProject/BasicOpenGLProject/RoomManager.h
@@ -5,7 +5,7 @@ class RoomManager {
 public:
 	static Room* currentRoom;
 	static int enemies;
-	static int score;
+	static float score;
 	RoomManager();
 	void draw() { RoomManager::currentRoom->draw(); };
 	bool check(float x, float y, Entity& Player) { return RoomManager::currentRoom->check(x, y, Player); };
